requirements.txt
Flask
tensorflow
Pillow
